refactor: drive maker runtime from events
